# perf: mesgdef micro optimisation