--- a/profile/mesgdef/cadence_zone_gen.go
+++ b/profile/mesgdef/cadence_zone_gen.go
@@ -62,10 +62,9 @@ func (m *CadenceZone) ToMesg(options *Options) proto.Message {
 
 	fac := options.Factory
 
-	arr := pool.Get().(*[255]proto.Field)
-	defer pool.Put(arr)
+	arr := pool.Get().(*[poolsize]proto.Field)
+	fields := arr[:0]
 
-	fields := arr[:0] // Create slice from array with zero len.
 	mesg := proto.Message{Num: typedef.MesgNumCadenceZone}
 
 	if uint16(m.MessageIndex) != basetype.Uint16Invalid {
@@ -78,7 +77,7 @@ func (m *CadenceZone) ToMesg(options *Options) proto.Message {
 		field.Value = proto.Uint8(m.HighValue)
 		fields = append(fields, field)
 	}
-	if m.Name != basetype.StringInvalid && m.Name != "" {
+	if m.Name != basetype.StringInvalid {
 		field := fac.CreateField(mesg.Num, 1)
 		field.Value = proto.String(m.Name)
 		fields = append(fields, field)
@@ -86,6 +85,7 @@ func (m *CadenceZone) ToMesg(options *Options) proto.Message {
 
 	mesg.Fields = make([]proto.Field, len(fields))
 	copy(mesg.Fields, fields)
+	pool.Put(arr)
 
 	mesg.DeveloperFields = m.DeveloperFields