--- a/profile/mesgdef/camera_event_gen.go
+++ b/profile/mesgdef/camera_event_gen.go
@@ -68,15 +68,14 @@ func (m *CameraEvent) ToMesg(options *Options) proto.Message {
 
 	fac := options.Factory
 
-	arr := pool.Get().(*[255]proto.Field)
-	defer pool.Put(arr)
+	arr := pool.Get().(*[poolsize]proto.Field)
+	fields := arr[:0]
 
-	fields := arr[:0] // Create slice from array with zero len.
 	mesg := proto.Message{Num: typedef.MesgNumCameraEvent}
 
-	if datetime.ToUint32(m.Timestamp) != basetype.Uint32Invalid {
+	if !m.Timestamp.Before(datetime.Epoch()) {
 		field := fac.CreateField(mesg.Num, 253)
-		field.Value = proto.Uint32(datetime.ToUint32(m.Timestamp))
+		field.Value = proto.Uint32(uint32(m.Timestamp.Sub(datetime.Epoch()).Seconds()))
 		fields = append(fields, field)
 	}
 	if m.TimestampMs != basetype.Uint16Invalid {
@@ -89,7 +88,7 @@ func (m *CameraEvent) ToMesg(options *Options) proto.Message {
 		field.Value = proto.Uint8(byte(m.CameraEventType))
 		fields = append(fields, field)
 	}
-	if m.CameraFileUuid != basetype.StringInvalid && m.CameraFileUuid != "" {
+	if m.CameraFileUuid != basetype.StringInvalid {
 		field := fac.CreateField(mesg.Num, 2)
 		field.Value = proto.String(m.CameraFileUuid)
 		fields = append(fields, field)
@@ -102,6 +101,7 @@ func (m *CameraEvent) ToMesg(options *Options) proto.Message {
 
 	mesg.Fields = make([]proto.Field, len(fields))
 	copy(mesg.Fields, fields)
+	pool.Put(arr)
 
 	mesg.DeveloperFields = m.DeveloperFields