--- a/profile/mesgdef/connectivity_gen.go
+++ b/profile/mesgdef/connectivity_gen.go
@@ -82,10 +82,9 @@ func (m *Connectivity) ToMesg(options *Options) proto.Message {
 
 	fac := options.Factory
 
-	arr := pool.Get().(*[255]proto.Field)
-	defer pool.Put(arr)
+	arr := pool.Get().(*[poolsize]proto.Field)
+	fields := arr[:0]
 
-	fields := arr[:0] // Create slice from array with zero len.
 	mesg := proto.Message{Num: typedef.MesgNumConnectivity}
 
 	if m.BluetoothEnabled != false {
@@ -103,7 +102,7 @@ func (m *Connectivity) ToMesg(options *Options) proto.Message {
 		field.Value = proto.Bool(m.AntEnabled)
 		fields = append(fields, field)
 	}
-	if m.Name != basetype.StringInvalid && m.Name != "" {
+	if m.Name != basetype.StringInvalid {
 		field := fac.CreateField(mesg.Num, 3)
 		field.Value = proto.String(m.Name)
 		fields = append(fields, field)
@@ -156,6 +155,7 @@ func (m *Connectivity) ToMesg(options *Options) proto.Message {
 
 	mesg.Fields = make([]proto.Field, len(fields))
 	copy(mesg.Fields, fields)
+	pool.Put(arr)
 
 	mesg.DeveloperFields = m.DeveloperFields