--- a/profile/mesgdef/course_gen.go
+++ b/profile/mesgdef/course_gen.go
@@ -64,10 +64,9 @@ func (m *Course) ToMesg(options *Options) proto.Message {
 
 	fac := options.Factory
 
-	arr := pool.Get().(*[255]proto.Field)
-	defer pool.Put(arr)
+	arr := pool.Get().(*[poolsize]proto.Field)
+	fields := arr[:0]
 
-	fields := arr[:0] // Create slice from array with zero len.
 	mesg := proto.Message{Num: typedef.MesgNumCourse}
 
 	if byte(m.Sport) != basetype.EnumInvalid {
@@ -75,7 +74,7 @@ func (m *Course) ToMesg(options *Options) proto.Message {
 		field.Value = proto.Uint8(byte(m.Sport))
 		fields = append(fields, field)
 	}
-	if m.Name != basetype.StringInvalid && m.Name != "" {
+	if m.Name != basetype.StringInvalid {
 		field := fac.CreateField(mesg.Num, 5)
 		field.Value = proto.String(m.Name)
 		fields = append(fields, field)
@@ -93,6 +92,7 @@ func (m *Course) ToMesg(options *Options) proto.Message {
 
 	mesg.Fields = make([]proto.Field, len(fields))
 	copy(mesg.Fields, fields)
+	pool.Put(arr)
 
 	mesg.DeveloperFields = m.DeveloperFields