… internal/cmd/fitgen/profile/mesgdef/mesgdef.tmpl
… profile/mesgdef/device_settings_gen.go
… profile/mesgdef/dive_alarm_gen.go
… profile/mesgdef/dive_apnea_alarm_gen.go
… profile/mesgdef/dive_gas_gen.go
… profile/mesgdef/dive_settings_gen.go
… profile/mesgdef/dive_summary_gen.go
… profile/mesgdef/event_gen.go
… profile/mesgdef/exd_data_concept_configuration_gen.go
… profile/mesgdef/exd_data_field_configuration_gen.go
… profile/mesgdef/exd_screen_configuration_gen.go
… profile/mesgdef/exercise_title_gen.go
… profile/mesgdef/field_capabilities_gen.go
… profile/mesgdef/field_description_gen.go
… profile/mesgdef/file_capabilities_gen.go
… profile/mesgdef/file_creator_gen.go
… profile/mesgdef/file_id_gen.go
… profile/mesgdef/goal_gen.go
… profile/mesgdef/gps_metadata_gen.go
… profile/mesgdef/gyroscope_data_gen.go
… profile/mesgdef/hr_gen.go
… profile/mesgdef/hr_zone_gen.go
… profile/mesgdef/hrm_profile_gen.go
… profile/mesgdef/hrv_gen.go
… profile/mesgdef/hrv_status_summary_gen.go
… profile/mesgdef/hrv_value_gen.go
… profile/mesgdef/hsa_accelerometer_data_gen.go
… profile/mesgdef/hsa_body_battery_data_gen.go
… profile/mesgdef/hsa_configuration_data_gen.go
… profile/mesgdef/hsa_event_gen.go