- `sync.Pool` now use max fields in the entire messages instead of 255. In the current Profilx.xlsx, the maximum fields is 156 (from session message), reducing approx. `99 * unsafe.Sizeof(Field{})` memory usage per created pool object.
- `ToMesg` now compare the invalid value of time without converting it to uint32 first. If the value valid, convert the value directly without using `datetime.ToUint32` since under the hood it will do comparison.

Benchmark:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/profile/mesgdef
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
               │  old.txt   │              new.txt               │
               │   sec/op    │   sec/op     vs base                │
RecordToMesg-4   214.1n ± 1%   174.6n ± 2%  -18.43% (p=0.000 n=10)


```